day4/mcp/requirements.txt
fastmcp
langchain-mcp-adapters
langchain
langchain-groq
python-dotenv
